+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- move log timestamp and coloring into the formatter
+
 ### Deprecated
+
+- `mex.common.logging.echo` is deprecated in favor of `logging.info`
 
 ### Removed